+++ core/commands.lua
@@ -88,12 +88,18 @@ SlashCmdList["FILGER"] = function(cmd)
             frame.Backdrop:Hide()
             frame.text:Hide()
         end
+    elseif (arg1 == "reset") and (Config["Plugins"].Tracer) then
+        Filger:ResetZone()
+    elseif (arg1 == "tracer") and (Config["Plugins"].Tracer) then
+        Filger:TracerEnable()
     elseif (arg1 == "") or (arg1 == "help") then
         print(" ")
         print(STRING_COLOR:format("Filger Commands:"))
         print(STRING_COLOR:format("test:"), "Test Filger.")
         print(STRING_COLOR:format("show:"), "Display Frames.")
         print(STRING_COLOR:format("hide:"), "Hide Frames.")
+        print(STRING_COLOR:format("reset"), "Reset saved variables.")
+        print(STRING_COLOR:format("tracer"), "Start/Stop filger aura tracer.")
         print(" ")
     end
 end